+++ Projects/UnitTests/Source/TestFunctions.cpp
@@ -29,6 +29,24 @@ TEST_CASE("Test Functions", "[Functions]")
 		REQUIRE(script);
 	}
 
+	SECTION("Test no return assignment function")
+	{
+		static const char * scriptText =
+			u8R"(
+	
+			function do nothing
+				-- do nothing
+			end
+
+			set a to do nothing
+		
+			)";
+
+		auto script = TestExecuteScript(scriptText);
+		REQUIRE(script);
+		REQUIRE(script->GetVariable("a") == nullptr);
+	}
+
 	SECTION("Test alternate names function")
 	{
 		static const char * scriptText =
@@ -99,11 +117,11 @@ TEST_CASE("Test Functions", "[Functions]")
 		static const char * scriptText =
 			u8R"(
 	
-			function return some constant integer
+			function some constant integer
 				return 42
 			end
 
-			function return some string val
+			function some string val
 				return "some string"
 			end
 
@@ -118,101 +136,12 @@ TEST_CASE("Test Functions", "[Functions]")
 		REQUIRE(script->GetVariable("b") == "some string");
 	}
 
-	SECTION("Test return value validation #1")
-	{
-		static const char * scriptText =
-			u8R"(
-	
-			function return somefunc
-				if true
-					return "some string"
-				else
-					return "some string"
-				end
-			end
-
-			set a to somefunc
-
-			)";
-
-		auto script = TestExecuteScript(scriptText);
-		REQUIRE(script);
-		REQUIRE(script->GetVariable("a") == "some string");
-	}
-
-
-	SECTION("Test return value validation #2")
-	{
-		static const char * scriptText =
-			u8R"(
-	
-			function return somefunc
-				if true
-					return "some string"
-				else
-				end
-				return "some string"
-			end
-
-			set a to somefunc
-
-			)";
-
-		auto script = TestExecuteScript(scriptText);
-		REQUIRE(script);
-		REQUIRE(script->GetVariable("a") == "some string");
-	}
-
-	SECTION("Test return value validation #3")
-	{
-		static const char * scriptText =
-			u8R"(
-	
-			function return somefunc
-				if (true)
-				else
-					return "some string"
-				end
-				return "some string"
-			end
-
-			set a to somefunc
-
-			)";
-
-		auto script = TestExecuteScript(scriptText);
-		REQUIRE(script);
-		REQUIRE(script->GetVariable("a") == "some string");
-	}
-
-	SECTION("Test return value validation #4")
-	{
-		static const char * scriptText =
-			u8R"(
-	
-			function return somefunc
-				if (true)
-				else if (false)
-					return "some string"
-				end
-				return "some string"
-			end
-
-			set a to somefunc
-
-			)";
-
-		auto script = TestExecuteScript(scriptText);
-		REQUIRE(script);
-		REQUIRE(script->GetVariable("a") == "some string");
-	}
-
 	SECTION("Test simple parameters")
 	{
 		static const char * scriptText =
 			u8R"(
 	
-			function return {a} plus {b}  
+			function {a} plus {b}  
 				return a + b
 			end
 
@@ -231,7 +160,7 @@ TEST_CASE("Test Functions", "[Functions]")
 		static const char * scriptText =
 			u8R"(
 	
-			function return {var a} plus {var b}  
+			function {var a} plus {var b}  
 				return var a + var b
 			end
 
@@ -252,7 +181,7 @@ TEST_CASE("Test Functions", "[Functions]")
 	
 			set public readonly x to 1
 
-			function return {a} plus {b}  
+			function {a} plus {b}  
 				return a + b
 			end
 
@@ -272,7 +201,7 @@ TEST_CASE("Test Functions", "[Functions]")
 	
 			set public readonly x x to 1
 
-			function return {a} plus {b}  
+			function {a} plus {b}  
 				return a + b
 			end
 
@@ -290,7 +219,7 @@ TEST_CASE("Test Functions", "[Functions]")
 		static const char * scriptText =
 			u8R"(
 	
-			function return {a} minus {b}  
+			function {a} minus {b}  
 				return a - b
 			end
 	 
@@ -312,7 +241,7 @@ TEST_CASE("Test Functions", "[Functions]")
 		static const char * scriptText =
 			u8R"(
 	
-			function return factorial { x }
+			function factorial { x }
 				if x <= 1
 					return 1
 				end
@@ -334,11 +263,11 @@ TEST_CASE("Test Functions", "[Functions]")
 			u8R"(
 	
 			-- Functions can be made up of keywords
-			function return loop function while
+			function loop function while
 				return 42
 			end
 
-			function return begin to while
+			function begin to while
 				return 99
 			end
 	 
@@ -358,11 +287,11 @@ TEST_CASE("Test Functions", "[Functions]")
 		static const char * scriptText =
 			u8R"(
 	
-			function return convert {string x}
+			function convert {string x}
 				return x
 			end
 
-			function return convertback {integer x}
+			function convertback {integer x}
 				return x
 			end
 
@@ -391,11 +320,11 @@ TEST_CASE("Test Functions", "[Functions]")
 			u8R"(
 	
 			-- Thto local function should be called from both scripts
-			function return localfunc
+			function localfunc
 				return 123
 			end
 
-			private function return privatefunc
+			private function privatefunc
 				return localfunc
 			end
 			
@@ -408,7 +337,7 @@ TEST_CASE("Test Functions", "[Functions]")
 	
 			-- This function should not be called, even though it
 			-- has the same name.
-			function return localfunc
+			function localfunc
 				return 456
 			end