+++ Source/JxLibrary.h
@@ -19,7 +19,7 @@ namespace Jinx
 		Library(RuntimeWPtr runtime, const String & name);
 
 		// ILibrary interface
-		bool RegisterFunction(Visibility visibility, ReturnValue returnValue, std::initializer_list<String> name, FunctionCallback function) override;
+		bool RegisterFunction(Visibility visibility, std::initializer_list<String> name, FunctionCallback function) override;
 		bool RegisterProperty(Visibility visibility, Access access, const String & name, const Variant & value) override;
 		Variant GetProperty(const String & name) const override;
 		void SetProperty(const String & name, const Variant & value) override;
@@ -36,7 +36,7 @@ namespace Jinx
 	private:
 		
 		// Private internal functions
-		FunctionSignature CreateFunctionSignature(bool publicScope, bool returnValue, std::initializer_list<String> name) const;
+		FunctionSignature CreateFunctionSignature(bool publicScope, std::initializer_list<String> name) const;
 		bool RegisterPropertyNameInternal(const PropertyName & propertyName, bool checkForDuplicates);
 
 		typedef std::map <String, PropertyName, std::less<String>, Allocator<std::pair<String, PropertyName>>> PropertyNameTable;